moved mediaSelection to interactionUtils and now display randomised genres
